Clarifies TLEN definition
States clearly that TLEN should be length (i.e. pos.right-pos.left+1) rather than subtraction 
(i.e. pos.right-pos.left) -- this was somewhat ambiguous before. 

Makes it clear that the sign character is prepended to the 

States that TLEN may be set to 0 whenever the information is unavailable, 
and gives examples as to when this would be the case.