Renames TLEN to ISIZE
Represents the current use of TLEN/ISIZE by declaring it to be an 
implementation specific integer with meaning only within the context 
of a particular application (if you get a SAM/BAM file today, you 
really do not know what is meant by the TLEN field without knowing 
which application has generated it). 

The value 0 is recommended for applications that do not use this 
field, which is backwards compatible with TLEN=0 meaning 
unknown/undefined template length. 

Also adds an extended tag 'IS' which can be used by future applications 
to indicate (in string form) which semantics apply to the field.