Commented out the SD and BD slice header aux tag definitions.
These tags are an accurate but insufficiently complete description of
the current implementations.  We need to more fully explore the corner
cases surrounding sequence and quality "*" so we can tighten up the
documentation.

Once done these can appear in (e.g. in v3.1) in a backwards compatible
manner, much as adding an additional aux tag to SAM works.