Clarify that only @HD/SQ/RG/PG/CO are valid headers
The intention has always been that the only valid header type codes are
those specified in this table, and reference implementations enforce or
take advantage of that.  Clarify that, and conversely clarify that header
field tags may be invented similarly to alignment aux tags.

This could be changed in future to add some new header types or to allow
user-defined header types, but this commit updates the spec to be clear
about current practice.

(Also fix the cline between @HD-SO and -GO.)