Clarify meaning of 0x40/0x80 FIRST/LAST flag bits
These reflect the physical sequencing technology rather than being
related to e.g. the ordering determined by POS fields after mapping.
I haven't explicitly mentioned that this is not "the segment with the
left-most mapped base" or similar as a full description accounting for
the ends being mapped to different chromosomes etc seems like it would
be more confusing, not less; but YMMV.  Fixes #169.