Generate PDFs into a new/ subdirectory
We will soon merge the gh-pages branch into this branch, and store
the canonical PDF specifications here alongside their .tex sources.

For convenient Git use when switching between feature branches etc,
newly generated PDFs are written to new/*.pdf and do not overwrite
the checked-in PDFs in the repository's root directory.

To generate development PDFs use "make" or e.g. "make new/SAMv1.pdf".
When generating a final PDF to be checked-in, copy it from the new
subdirectory, with "mv"/"cp" or via e.g. "make SAMv1.pdf".

(If a document uses a .ver file, it will be a good idea to check in
the .pdf in a *separate* commit after committing changes to the .tex
source, so that the commit id from .ver includes the corresponding
source changes.)