Add @SQ-AN alternative reference sequence names (PR #220)
Enables tools to allow users to make queries with e.g. "1" or "chr1"
interchangeably.  Also allows for the possibility of tools using an alias
when displaying sequence names to the user.  Hat tip @lindenb, fixes #100.

However aliases must not appear elsewhere within the SAM file, in
particular not in RNAME/RNEXT fields.  This ensures that files will
still be parsed correctly by non-@SQ-AN-aware tools.