htsget protocol identifier in JSON text (#230)
Add protocol identifier to htsget API response.

Make all responses from the htsget API a JSON object which
contains the single key 'htsget', and the response data in a JSON
object as the value.  This allows clients using the API to verify
after receiving the first few bytes that the response is an htsget
message and not some other JSON-based protocol, or a different
format altogether.