Add BAM version 2
The magic number is changed to "BAM\2" to allow software to detect
the new version.

The size of `n_cigar_op` is increased to uint32_t.

An extra field `bam2_hdr_flags` is added to the header, and
`bam2_flags` is added to the alignment records.  Both of these fields
are reserved for future enhancements.

The obsolete `bin` field is removed from version 2 files.

A new `BV` tag is added to the SAM `@HD` line as a (rather ugly)
way of hinting which BAM version should be used.