Revise @SQ M5 description following feedback comments
The revised description is too long for a table cell, so it gets
put into its own section.

Make the description of how the digest is calculated clearer.

Expand on padding characters a bit, including a link to the section
about padded SAM.

Add examples, including one that uses pad characters.