Add POST method to htsget
Query parameters are passed as JSON.  It is possible to request
more than one location in the returned data.  The region list acts
as a filter on the requested file, so alignment records are
only returned once, even if they overlap more than one region.

Adds 413 to list of htsget error codes so servers can limit the
size of POST request they support.

To keep things simple, URL query parameters are not allowed in
POST requests.