Rewrite auxiliary field section to really show the BAM representation
Whether the array element _count_ is properly uint32_t or int32_t has
not been described previously. Ideally it would be unsigned (as it is
an item count!), but it is limited by block_size which the spec already
describes as a signed int32_t, and it is implemented as such in HTSlib.

Fixes #273.