Fix BAI linear index description: "overlapping", not "starting in"
As per Heng Li, "Re: Understanding the linear index", on the
samtools-help mailing list in June 2010 [1]:

   "The spec is wrong here. We should store the file offset of the
    leftmost alignment that *overlaps* (not starts in) the 16kb window.
    The implementation is different from what the spec is saying."

Also hat tip to Adam Birnbaum of Edico Genome who rediscovered this
in 2015 (private communication).

Rediscovered once again in 2016 in samtools/htslib#405. As noted in
a comment on that issue [2], samtools, HTSlib, and Picard have always
constructed the linear index based on *overlapping* alignments.

[1] https://sourceforge.net/p/samtools/mailman/message/25602941/
[2] https://github.com/samtools/htslib/issues/405#issuecomment-243435508