Bug fix NP encoding type.
It is defined as 'int' in the main table, but section 1.5 lists the
encoding type as long.  This is both pointless (the mate's AP field is
int, not long) and also not how it is implemented in both htsjdk and
htslib.  The origin of this stems all the way back to the CRAM 1.0
word document, but was apparently missed when fixing the spec to match
the initial Java implementation.