Version 0.2 of the refget API as discussed in PR #327
Text as exported from previous Google Docs document:
https://docs.google.com/document/d/1q2ZE9YewJTpaqQg82Nrz_jVy8KsDpKoG1T8RvCAAsbI
as of 27th July 2018 (last edit was made 25th June 2018). Format was
changed into Markdown with HTML.

This is version 0.2 of the refget specification. Refget supports the
retrieval of sequences using an identifier generated by a hash digest of
the full length underlying sequence. It also supports retrieval of
metadata concerning this identifier including sequence length and
aliases known to the server.