Clarifying circular start authority
Circular chromosomes are a potential issue as to create the checksum it must have a start and an end i.e. a break-point in the circular DNA. This is out of scope for this specification and we expect other authorities to have made that call for us i.e. a submitter to an INSDC archive will have to make this call.