Fix type of GQ in example (#299)
Issue #263 pointed out that the example structural variant VCF includes a GQ field which is typed as a Float. It's defined elsewhere as an Integer.

This fixes it in the 4.1 -> 4.3 documents. It also updates the example vcf header version number to match the spec version it's an example for.

Fixes #263