Rerun LaTeX by checking .aux/etc files rather than scraping .log
\RerunFileCheck does not work with --output-directory, at least for
foo.out (it looks at the unchangingly non-existent ./foo.out rather
than the changing new/foo.out). This means that some PDFs have been
missing their PDF Tables of Contents as LaTeX has not been rerun
often enough.

Instead of grepping for \RerunFileCheck's messages, use a new script
that checksums (using POSIX cksum rather than md5sum, which has a
different command name on macOS) LaTeX's state files (.aux/.out/.toc)
and iterates until fixed point. This is similar to a simple form of
latexmk (which we don't want to use as it may not be installed).

None of the documents actually have indices, so remove \makeindex
from the .tex files it appears in. This reduces the number of LaTeX
runs needed for BCFv2_qref and CSIv1. Don't ignore or clean .idx,
as if some index reappears we'll want to know about it to add code
to invoke makeindex etc.