Support table of contents in refget. See issue #341. (#343)
During refget's PRC, the reviewers suggested adding a table of contents
to the specification. Kramdown can support this via the {:toc} tag.
Anything that should not be included in the TOC is tagged
with {:.no_toc}.

The config is setup to build the TOC for all h1 and h3 tags, which
on first inspection looks sufficient.