[SQUASH ME] Suggested rewrite without bullet list
Use "original alignment entry" to be more specific.
Say "textual SAM representations" (but "generally" because STRAND is not
direct from a SAM field); now there's nothing needed for some subfields
so compress the bullet list into a list in a sentence.

Restore useful text in OP even if we're deprecating it.