Improve positioning of "Reserved INFO/genotype keys" tables
Compress the list of columns and encourage a pagebreak in "4 REF" to get
more vertical space for Table 1 on page 8. Use \longtable so Table 1 can
be broken across pages; don't use \begin{table}, so that the tables are
positioned within their related ("8 INFO" / "Genotype fields") text.
Keep initial "INFO keys used for structural variants" paragraphs together
so they don't creep back to page 11.

(Previously it was very easy to misread the INFO keys table as pertaining
to the genotype text surrounding it.)