Clarify NM and MD creation in CRAM (PR#463)
Fixes - mostly - #453.

It doesn't cover the nuances I want for CRAM 4.0, but that will be a separate
commit.  It also glosses over the difference between embedded references and
external references.  We not currently state whether they must be identical
and this is a difference in the htslib and htsjdk implementations so it is
left undecided for now.