Show double quotes around ##ALT Description
Unlike for the other headers, the section on ##ALT does not show the Description field as having double quotes around it.

However §1.4.2 (Information field format) says Description must be surrounded by double-quotes and this can be construed as applying to Descriptions in all headers, all the other headers show it, all the examples of ##ALT in the specification have double quotes, and tools such as bcftools and GATK use double quotes in the ##ALT fields they output. So this is presumably just an omission in the spec.

Hat tip @lltw; see pysam-developers/pysam#881.