Further CRAMv4.0 updates
This includes a massive reorganisation of the document, which is long
overdue.

It also now replaces EXTERNAL with a variety of type specific codecs
for signed, unsigned and bytes; ditches HUFFMAN (along with all other
bit codecs) and adds CONST codecs instead; removes the CORE block;
removes "external" for block descriptions, instead favouring data
blocks as a way to distinguish from header / structure containing
blocks.

As yet, nothing implements this variant of V4.0! However it's far
cleaner to describe and hopefully far simpler to implement due to
the aggressive culling of largely unused features.