Add MD string definition alongside the motivating example (PR #500)
Fixes #485. Behaviour for the more exotic CIGAR operators derived
from that of the `samtools calmd` command and (equivalently) the
MD computation in HTSlib's cram_encode.c.

This documents the status quo, which includes '0' lengths separating
any otherwise-adjacent distinct base strings. (TODO) We may add a
footnote or similar acknowledging the possibility of relaxing this.