Correct CRAM MF bit 2: it means unmapped, not mapped. (PR #511)
The pseudocode in CRAMv3.tex was already correct.

This flaw appears to date all the way back to the initial CRAM1.0 pdf
document:

https://www.ebi.ac.uk/sites/ebi.ac.uk/files/groups/ena/documents/cram_format_1.0.1.pdf

Implementation-wise, the initial commit for the CRAMtools Java
implementation of CRAM (enasequence/cramtools#6576802) had (Oct 28
2012):

        public byte getMateFlags() {
                if (mateFlags == null) {
                        byte b = 0;
                        b |= mateMapped ? 1 : 0;
                        b <<= 1;
                        b |= mateNegativeStrand ? 1 : 0;
                        mateFlags = new Byte(b);
                }
                return mateFlags;
        }

7 days later, commit enasequence/cramtools#f0cf055 changed that with:

        public byte getMateFlags() {
                if (mateFlags == null) {
                        byte b = 0;
-                       b |= mateMapped ? 1 : 0;
+                       b |= mateUmapped ? 1 : 0;
                        b <<= 1;
                        b |= mateNegativeStrand ? 1 : 0;
                        mateFlags = new Byte(b);

This is shortly before the first release of CRAM (indeed the commit
message for above is "cram2bam, almost there").  It appears the spec
was never updated so has been wrong since day 1.

Also minor tidy up to wording of BF flag 0x20.