SAM: explicitly limit the TLEN field to the "primary alignment" (PR #522)
PNEXT and RNEXT are explicitly stated to point to the next primary
alignment.  So for example a secondary alignment of read 1 will
point to the primary alignment of read 2.

TLEN however had no such clause, so if read 1 and read 2 both align as
a valid pair to chr1 (primary), with a secondary alignment of read 1
also reported on chr2, then it was technically still not permitted to
fill out TLEN for any read, including the primary pair, due to the
existence of that secondary alignment.

Co-authored-by: John Marshall <jmarshall@hey.com>