Clarify structured vs unstructured meta-information lines
Introduce the term "unstructured meta-information line", and reword this
section so it describes the two flavours of meta-information line clearly.
Specify that an unstructured value must not start with `<` (so that
structured/unstructured are easily distinguished) and must be non-empty.

Remove `<>` from unstructured `##pedigreeDB` example. PR #88 removed
the `<>` from one instance of `##pedigreeDB=<url>` presumably on the
grounds that they were merely metasyntactic variable notation and not
intended to appear literally, but missed this instance.