Move service-info to top level
Submitted as a separate commit to enable backing out the change if required. @tcezard suggested on PR #479 "Shouldn't we take the opportunity of a major version change to move the service-info at the top level". This commit moves service-info to the top level of the implementation. It is assumed that namespacing of the endpoint will happen elsewhere