Add GitHub Actions to build PDFs and comment on pull requests
Use GitHub's recommended two-part approach [1][2] to separate building
the PDFs (which runs in the PR branch's context and has only a read-only
GH token) from commenting on the PR (which runs on the master branch and
has write access to the repository).

Add a brief README (to be expanded as this beds in), and mention in the
old README that the old implementation will soon be removed: the CircleCI
web hook and its access token have been removed from the hts-specs repo.

[1] https://securitylab.github.com/research/github-actions-preventing-pwn-requests/
[2] https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#using-data-from-the-triggering-workflow