Clarify description of structured vs unstructured meta-information lines (#620)
* Clarify structured vs unstructured meta-information lines

Introduce the term "unstructured meta-information line", and reword this
section so it describes the two flavours of meta-information line clearly.
Specify that an unstructured value must not start with `<` (so that
structured/unstructured are easily distinguished) and must be non-empty.

Remove `<>` from unstructured `##pedigreeDB` example. PR #88 removed
the `<>` from one instance of `##pedigreeDB=<url>` presumably on the
grounds that they were merely metasyntactic variable notation and not
intended to appear literally, but missed this instance.

* Remove `<>` from VCF test file `##pedigreeDB` lines

The specification now consistently reflects that `##pedigreeDB`'s value
should not be delimited by angle brackets (despite being an URL!).

Adjust the failed_meta_pedigreedb_002.vcf files as the claimed cause
of failure is the invalid URL hostname rather than the angle brackets.

* Fix misleading `##pedigreeDB=<url>` notation in older VCF specifications

PR #88 removed the `<>` from `##pedigreeDB=<url>` in VCFv4.3.tex,
presumably on the grounds that they were merely metasyntactic variable
notation and not intended to appear literally. As some readers still
refer to these older documents, remove the misleading notation here too.

* Mention structured lines that are not defined by the VCF specification