Clarify SAM file encoding (ASCII, UTF-8 in designated fields) (#670)
Reword so that it is clear that the encodings specified apply to the
*entirety* of SAM file contents. Mention allowable line terminators,
and note that most UTF-8 line terminating characters are invalid
(as this whitespace is in the ASCII-only parts outwith UTF-8 fields).
Fixes #664.

Spell out that the locale considerations are about e.g. requiring that
floating-point values use '.' rather than a localised decimal point.