Document int32_t SAM header length in CRAM header (PR #683)
The first 4 bytes of the block holding the SAM header is a length
field, for the header text itself.  This seems redundant, but it has
been this way since at least 1.1.

Note this field is signed.  I do not know the reasoning, but it is
perhaps originates in the lack of unsigned types in the original Java
implementation.

(For CRAMv4, this 4-byte preamble should be completely removed and the
length dictated yb the surrounding block structure instead.)

Fixes #678