Set safe.directory within the container-based GH Action step
The LaTeX step that does the pr-pdfs workflow's actual PDF generation
runs as our own GitHub Action, so runs within a Docker container and hence
as a different Unix user from the uid that owns the checked-out files.

As described in actions/checkout#766, such actions need to work around
recent Git's directory ownership security check by marking the repository
directory as a Git safe.directory. (IMHO this is something that should
be taken care of by the GH runner's setup for container-based steps.)

Make a separate run.sh entrypoint script so we can add this easily,
and take advantage of the new script to improve the step's logging.