Add CIGAR =/X entry to version history appendix
These CIGAR operations were added in 07dc1c67a717 in July 2010,
which was the initial addition of the TeX specification; they
were not present in the previous Pages document. It's so long ago
as to be barely relevant now, but it's worth mentioning them as
requiring VN:1.3 rather than VN:1.0.

In the SAM regexp in §1.4, write the operations in the familiar
canonical order (though it doesn't affect the meaning of the regexp).
Define \cigarops{...} to improve the formatting of lists of CIGAR
operations like "M/I/D" by making the slashes non-\tt, and also use
this in SAMtags.tex.