Explicitly permit eg N+m in MM tag (PR#799)
The text already states that an unmodified base of N means we count
any base type, but base N code N in the table is a little misleading
as to the intention.  It was intended to mean any unspecified
modification, in the same way C+C is any unspecified C mod, but in
this case it's against all bases rather than a specific base type.

However that doesn't solve the issue of whether we can record specific
mods against any "fundamental" source base.  Clarified this by adding
an extra line to the table and some text.  (However note this doesn't
necessarily imply downstream processing tools will not do any
compatibility assessment and reject N+m when the SEQ base is a T.)

Fixes #785