Add a footnote on case-insensitivity of RG PL (PR #684)
This is not changing what is valid / permitted, and indeed this
hopefully clarifies it further.  However the practicality of dealing
with wide-spread non-compliant data with lowercase PL values is that
tools may wish to be lenient and use case-insensitive matching.

Also removes test/sam/failed/hdr.RG6.sam due to explicitly testing
against the use of lower-case PL fields.  While strictly not
conforming, it's overly harsh if we are advocating a more
spec-tolerant testing regime for PL.

Fixes #679.