Copy the tag type text from SAM spec to SAMtags (PR #804)
SAM section 1.5 clearly defines the standard SAM tag types along with the
expanded codes used in the B byte-array type.  This text has been
copied into the SAMtags document to remove a rather woolly definition
there.  The text describing lower-case tags has been removed, because
this was already discussed in further detail at the end of the SAMtags
document (and is itself somewhat woolly when it comes to half-upper
half-lower combinations due to the addition of draft tags).

Fixes #798