feat: add tags related to optical duplicates.
Picard has long been able to add Sam Tags that provide information about duplicate templates: Their type (Library or Sequencing), the size of the sets they are part of, and the queryname of the representative template within the set that is _not_ a duplicate. As I look to use these tags in another repository (fgbio) I wanted to make sure that the tags are generally accepted and well defined.