feat: add proof of concept solution for #814
I added all the SamTags and many of the SamHeader tags as well as three Enums and the BAM magic string into three different yaml files. I then ran reconstant on said files and obtained the autogenerated code.

reconstant doesn't currently have an R nor a LaTeX output mode, but its a simple enough code that it could be injested and modified into this code-base, or I could submit a PR and we could continue using from it's current location.

This PR is meant to provide something to discuss in the next meeting, it's not ready for merging.