Rework the CRAM bit encoding example and clarify text
Given it's working in bits, the example is much clearer if we describe
the bits instead of hex, especially distinguishing bits set/unset from
bits-yet-to-use.

I also reworked the note at the end of the section as it was quite
hard to follow.  I gave it real examples of BETA and HUFFMAN to
clarify what is meant by the decoder needing to know the number of
bits to consume.

Fixes #812.